+++ CHANGELOG.md
@@ -33,7 +33,8 @@ Changelog
  * added `Puli::disablePlugins()`
  * added `Puli::arePluginsEnabled()`
  * added `Puli::isStarted()`
- * added support for extra keys in `PackageFile`
+ * added support for extra keys to `PackageFile`
+ * added support for extra keys to `RootPackageFileManager`
 
 * 1.0.0-beta2 (2015-01-27)